# sources/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация по API",
        "description": "API документация",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/vacancy/destroy": {
            "delete": {
                "tags": [
                    "Vacancy"
                ],
                "summary": "Удаление вакансии",
                "description": "Возвращение JSON объекта",
                "operationId": "386804a46c923cc185e2197f95c2db59",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Ключ вакансии",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "e8ff11fa-822b-11f0-8411-10f60a82b815"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "УСПЕХ!"
                    }
                }
            }
        },
...
        "/api/v1/vacancy/read": {
            "get": {
                "tags": [
                    "Vacancy"
                ],
                "summary": "Получение вакансии",
                "description": "Возвращение JSON объекта",
                "operationId": "4e1249a7f9f37071a553469a56dfab9f",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Ключ вакансии",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "e8ff11fa-822b-11f0-8411-10f60a82b815"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "УСПЕХ!"
                    }
                }
            }
        },
...
        "/api/v1/vacancy/update": {
            "post": {
                "tags": [
                    "Vacancy"
                ],
                "summary": "Обновление вакансии",
                "description": "Возвращение JSON объекта",
                "operationId": "8945c8510c407e8f73ce3e20c56d6677",
                "requestBody": {
                    "description": "Данные для обновления вакансии",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "key": {
                                        "type": "string",
                                        "example": "e8ff11fa-822b-11f0-8411-10f60a82b815"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Вакансия"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание вакансии"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "УСПЕХ!"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}